Fixed Issue #5 Added a flag for processing a single account: -a
Usage:  ./copyscript -s (source) -a username
or: ./copyscript -s (source) -a domain.com

This mode will be useful when wanted to focus on one account or are
migrating a custom list of accounts (loop the values in)
Additionally,  the domain or username can be used here.  This should
make it easier when customers are providing accounts to migrate by
their domain name.